Initial commit
This commit leaps the project forward to being able to ingest a provided
schema file, validate it and then build a simple form. The forms that
are able to be generated are basic and do not cover a lot of what you
can express through JSON schema.

This is built on top of fge/json-schema-validator which will be used for
validating incoming data. I have written a little Java to reuse some of
the internals of this for building a persistent map version of the
schema that has been reffed.