Allow setting a CSRF input field through fn
In order to protect from CSRF attacks a user of this library will want
to inject a form field into the 'rendered' html that the application
will check when receiving a POST. This allows a user to pass in a
function that will generate some hiccup/html to be injected into the
form.

I have augmented the test webapp to use ring-anti-forgery to demonstrate
this functionality.