Validate form parameters against json schema
We want to receive a map with the form fields as keys and any errors
they might have as the value. These errors should be keywords indicating
the kind of error, such that the application using this library can
choose what the error message should be.

I had to move the creation of the Schema back into Java so that I could
more easily manipulate the JsonSchemaFactory to add in the two new
keywords and the password format.

Because of the addition of these new fields we have deviated away
from the spec draft-04 schema so now files that are read by this library
should use `http://forms.beardandcode.com/draft-04/schema#` as their
schema value. This schema should mirror draft-04 but with the addition
of the extras for this library.

We ditch any form parameters that have an empty field, if we don't do
this then they pass the required check in the schema and this is not
what would be expected.