Do a deep check when validating
We want to do a deep check as this produces what a user expects. When
doing shallow, if you are missing required fields it will not check any
further so something you have filled in might still be wrong. You would
only find out if it was wrong when you had filled in everything
required. This seems silly, a deep check will fix this.