Render previous values and errors into forms
Once a user submits a form and has made a few errors in it, we should be
able to re-render the form with their old values and the problems that
they had in it. We never want to render passwords back to a user as this
is against the common standard.

We allow a developer consuming this library to provide their own
function for turning the validation error keywords back into text. This
should allow for specific variation when required.