Render errors correctly when using nested forms
Dropped the removal of empty maps as it meant the expected behaviour
when validating didn't happen. In the nested example, if no fields in
address were filled out it wouldn't say the postcode was required.
Doesn't feel right.

Switched error keys for pointers to the bit of the map, this should
allow of root errors. Whoop!