Load schema in clojure rather than ObjectMapper
When the schema exists in a JAR trying to coerce a URL to a File using
`(io/file)` causes an error say it is "not a file". This is because the
URL starts with `jar:` rather than `file:`. `slurp` handles reading from
a resource and ObjectMapper can process a String, so it seemed easier to
just use `slurp`.